requirements.txt
telethon==1.42.0
aiogram==3.13.1
prometheus-client==0.20.0
APScheduler==3.10.4
python-dotenv==1.0.1
asyncpg==0.29.0
